Roughly Finished PIC Slave Control Code
PIC code for both the DC motors is roughly finished. Haven't tested the
SPI communication, but we don't have anything to test it with right now.
As of right now, the PIC will read in the SPI input (8-bit character),
and parse out the speed and direction info, and send out values
accordingly.